% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-data.R
\name{get_bus_polygon_centroids}
\alias{get_bus_polygon_centroids}
\title{get_bus_polygon_centroids}
\usage{
get_bus_polygon_centroids(dat)
}
\arguments{
\item{dat}{An \pkg{osmdata} object representing either \code{key = "bus"} or
\code{key = "public_transport"}, both of which may return polygons
representing bus stations or platforms}
}
\value{
Modified verion of input in which centroids of polygons have been
added to the \code{$osm_points}.
}
\description{
get_bus_polygon_centroids
}
