% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-data.R
\name{get_who_busstops}
\alias{get_who_busstops}
\title{get_who_busstops}
\usage{
get_who_busstops(city = "kathmandu")
}
\arguments{
\item{city}{Name of city for which bus stops are to be obtained}
}
\value{
The \pkg{sf}-formatted data object (invisibly)
}
\description{
Extract OSM bus stops for given location (\code{city}), and save them in the
data directory. Note that coordinates only are extracted, because some bus
facilitites are polygonal stations which are simply converted to centroids.
}
